+++ src/main/java/io/seqware/pancancer/JobUtilities.java
@@ -298,10 +298,6 @@ public class JobUtilities {
                         + "-v "
                         + workflowDataDir
                         + ":/workflow_data "
-                        // link in the working directory for local file uploads as well, if needed
-                        + "-v "
-                        + DEWrapperWorkflow.SHARED_WORKSPACE_ABSOLUTE
-                        + ":/datastore "
                         // link in the pem key
                         + "-v "
                         + pemFile